Chore: build 문제 해결 - UserController 합침
