+++ src/main/java/com/dev/moim/domain/user/dto/ReviewDTO.java
@@ -1,4 +1,4 @@
-package com.dev.moim.domain.account.dto;
+package com.dev.moim.domain.user.dto;
 
 public record ReviewDTO(
         Long reviewId,
